+++ pkg/keymanagementprovider/keymanagementprovider.go
@@ -24,11 +24,9 @@ import (
 	"strings"
 	"sync"
 
-	corecrl "github.com/notaryproject/notation-core-go/revocation/crl"
 	"github.com/ratify-project/ratify/errors"
 	"github.com/ratify-project/ratify/internal/constants"
 	ctxUtils "github.com/ratify-project/ratify/internal/context"
-	"github.com/ratify-project/ratify/internal/logger"
 	vu "github.com/ratify-project/ratify/pkg/verifier/utils"
 	"github.com/sigstore/sigstore/pkg/cryptoutils"
 )
@@ -88,12 +86,6 @@ var certificateErrMap sync.Map
 //
 //	map["<namespace>/<name>"] = error
 var keyErrMap sync.Map
-
-// logOpt is a logger.Option configuration for the KeyManagementProvider component.
-// It sets the ComponentType to logger.KeyManagementProvider.
-var logOpt = logger.Option{
-	ComponentType: logger.KeyManagementProvider,
-}
 
 // DecodeCertificates decodes PEM-encoded bytes into an x509.Certificate chain.
 func DecodeCertificates(value []byte) ([]*x509.Certificate, error) {
@@ -248,32 +240,4 @@ func hasAccessToProvider(ctx context.Context, provider string) bool {
 		return !vu.IsNamespacedNamed(provider)
 	}
 	return strings.HasPrefix(provider, namespace+constants.NamespaceSeperator) || !vu.IsNamespacedNamed(provider)
-}
-
-// SupportCRL checks if the certificate supports CRL
-func SupportCRL(cert *x509.Certificate) bool {
-	return cert != nil && len(cert.CRLDistributionPoints) > 0
-}
-
-// cacheCRL caches the Certificate Revocation Lists (CRLs) for the given certificates using the provided CRL fetcher.
-// It logs a warning if fetching the CRL fails but does not return an error to ensure the process is not blocked.
-func CacheCRL(ctx context.Context, certs []*x509.Certificate, crlFetcher corecrl.Fetcher) {
-	var wg sync.WaitGroup
-	for _, cert := range certs {
-		// check if the certificate supports CRL
-		if !SupportCRL(cert) {
-			continue
-		}
-		for _, crlURL := range cert.CRLDistributionPoints {
-			wg.Add(1)
-			go func(url string) {
-				defer wg.Done()
-				if _, err := crlFetcher.Fetch(ctx, url); err != nil {
-					// Log error but do not return. Ensure unblock on CRL download failure
-					logger.GetLogger(ctx, logOpt).Warnf("failed to download CRL from %s: %v", url, err)
-				}
-			}(crlURL)
-		}
-	}
-	wg.Wait()
 }
